Copter: Moved commands_logic to mode_auto
Moved methods from commands_logic.cpp to
the mode_auto.cpp. Method order is
the same as declared in mode.h
